Modularizing codegen.
